+++ src/Select/Select.tsx
@@ -1,50 +1,13 @@
 import React, { useRef, useCallback } from 'react';
 import classNames from 'classnames';
 
-import FormControl, { FormControlProps } from '../FormControl';
+import FormControl from '../FormControl';
 import FormLabel from '../FormControl/FormLabel';
-import { OptionType, RawValueType } from './types';
+import { SelectProps } from './types';
 import Context from './context';
 import Listbox from './Listbox';
 import useSelect from './useSelect';
 import './styles/select.scss';
-
-export type SelectProps = {
-  /**
-   * 	The id of the select element.
-   */
-  id?: string;
-  /**
-   * Placeholder of select.
-   */
-  placeholder?: string;
-  /**
-   * Array of options that populate the select menu.
-   */
-  options: OptionType[];
-  /**
-   * Current selected option. Use when the component is controlled.
-   */
-  value?: RawValueType;
-  /**
-   * 	The default selected option. Use when the component is not controlled.
-   */
-  defaultValue?: RawValueType;
-  /**
-   * Name of the HTML Input (optional - without this, no input will be rendered)
-   */
-  name?: string;
-  /**
-   * The aria-label of the select element.
-   */
-  ariaLabel?: string;
-  /**
-   *	Callback function fired when an option is selected.
-   */
-  onChange?: (newValue: RawValueType) => void;
-  className?: string;
-  [propName: string]: unknown;
-} & Omit<FormControlProps, 'children' | 'htmlFor'>;
 
 const Select: React.FC<SelectProps> = ({
   label,
@@ -74,7 +37,7 @@ const Select: React.FC<SelectProps> = ({
     isExpanded,
     setIsExpanded,
     setSearch,
-  } = useSelect(optionsProp, id, value, defaultValue, onChange);
+  } = useSelect({ options: optionsProp, id, value, defaultValue, onChange });
 
   const timeOutId = useRef<number>();
   const refSelControl = useRef<HTMLButtonElement | null>(null);
